TODO: http://www.grammaticalframework.org/irc/freenode_%23gf_20161210.log http://www.grammaticalframework.org/irc/freenode_%23gf_20161212.log
ComplSlash( SlashV2V v {s=[]; a=Simul} {s =[]; p= CNeg False } vp ) np;
parses "tell the child to not eat the marshmallow" but not "tell the
child not to eat the marshmallow"
In connection with the empty strings,
https://groups.google.com/forum/#!topic/gf-dev/aRjt_2JfvA0
I don't understand. infVP, which SlashV2V calls, returns "not" ++ "to".
The CNeg I think infVP is getting must be not what it is getting. And it
is picking up the "not" from the vp's ad(verb) field and parsing it a a
split infinitive!